Create Money API
The first version of the 0.9 API

- Created a new money-api module that houses a Java Only API that has
  been revised from the current

- The new Java API is designed to be very flexible from an
  implementation perspective.  We have boiled as many pieces that we
could find that were specific to the original akka based implementation.
That left us with a very clean and simple API.  Some changes are
highlighted below

Some substantive changes include:

- Removed timeouts.  Timeouts are an artifact of the original akka
  design.  In that design, we completely decoupled the calling client
from the span implementation.  Each Span was an Actor.  If a Span Actor
never received a stop message, then it would live forever, leading to a
memory leak.  The new API does not have a bias toward
implementation, so a timeout is not necessary.

- Removed bias toward thread local. There are applications for money
  where thread local are not a good fit.  We heavily relied on thread
local for the original money implementation.

- Java only api.  This allows us to provide a java only implementation
  of money that can built without the need for scala.  An example would
be Android, where the size of the scala library would be prohibitive as
a dependency.