Add MetricRegistry factory
This patchset allows to configure a different way on how
MetricRegistry objects are created:
Up to now these were created within money itself with no exposition to
the rest of the system. That makes it impossible to register other
Dropwizzard reporters or exporters on it.
With this patchset it is now possible to provide a custom MetricRegistry
factory and do the required initialization of/with the MetricRegistry
object.
It is compatible to Java so even Java clients (using the Scala money)
can implement and configure a custom factory.