Improve MetricRegistryFactory configuration
The current appraoch requires two properties to be set to enable a
custom MetricRegistryFactory:

```
  metrics-registry {
    class-name = "com.acme.CustomMetricRegistryFactory"
  }

  emitter {
    emitters = [
      {
        name = "span-metrics-emitter"
        class-name = "com.comcast.money.metrics.SpanMetricsCollector"
        subscriptions = [Trace]
        configuration = {
          metrics-registry {
            class-name = "com.acme.CustomMetricRegistryFactory"
          }
        }
      }
    ]
  }
```

With this change, the configuration simplifies down to

```
  metrics-registry {
     class-name = "com.acme.CustomMetricRegistryFactory"
     configuration = {
     }
  }
```

for the money system and providing a custom 'configuration' section for the factory.

Also, as it is now an Akka extension, the factory is created only once for the whole system.