Update stop span to take a boolean
Instead of taking a Note.  This is part of refactoring to use the new
api.Note.  When I did the initial refactoring, a lot of inconsequential
code changes were necessary because the existing stopSpan method
required a Note.  By moving this to a boolean instead, I will reduce the
amount of code changes needed when I actually move to the new Note.