Added additional note factory methods
This is another change in advance of refactoring to use the api Note
instead of the existing core note.

There are several factory methods expected in the existing core code
that assume default values.  This PR brings parity to the new api Note
with the existing core Note classes.  This will drastically reduce the
number of changes needed when we do finally update to use the api Note.