Add configuration for the LogEmitter format
This change updates the LogEmitter to allow the format and keys to be
configured to allow customization of the log output.

All of the configuration values have a fallback to the current value,
leaving the current log format unchanged.

The template for logging individual values can be specified as
a log-template property.  The span-start string and null-value can also
be customized.

Each of the field keys can be cutomized with these properties:
  span-id = "span_id"
  trace-id = "trace_id"
  parent-id = "parent_id"
  span-name = "span_name"
  app-name = "app_name"
  start-time = "start_time"
  span-duration = "span_duration_ns"
  span-duration-ms = "span_duration"
  span-success = "span_success"

Also a new span-duration-ms field is added to allow logging the duration
in ms.  By default it is disabled, however the span-duration-ms-enabled
configuration value can be used to enable it.