Add money tracing support to AmazonWebServiceClient.
- AmazonWebServiceClient is what most aws-java-sdk platform services use internally.
DynamoDB client is a good example that makes use of this client.
- AmazonWebServiceClient encapsulates the actual instance of the HttpClient and does not
want its clients to interact with it directly. The only way you can do so is through RequestHandlers.
- This implements a request handler to AmazonWebServiceClient that sends the money trace data
to the platform as a HTTP request header. Once the platform service reads the trace data from
the client request object, it is supposed to span on that.
- This was tested just with the AmazonDynamoDBClient but it should work with any of the other
 AWS platform services that make use of the AmazonWebServiceClient internally.