Introducing Akka support
This is the first part of a PR to support Akka's actor system with money

The first step is to support a consistent storage for span's in the
context of actors exchanging messages.
This means spans has to be started and stopped manually as depicted in
the test case `MoneyExtensionSpec`.

Further PRs will add convinience methods to support automatic starting &
stopping for messsages flowing through the actor system